lifted items to app component
